better ilp tests
